keos

Module addressing

Physical and Virtual Memory Addressing Schemes.

This module provides abstractions for virtual address and physical address. In the keos kernel, kernel virtual memory is directly mapped to physical memory. The first page of kernel virtual memory maps to the first frame of physical memory, the second page maps to the second frame, and so on. This direct mapping allows the kernel to calculate the physical address from the kernel virtual address with simple arithmetic operations, by adding or subtracting a constant offset.

The module defines three primary types for memory addresses: Kva for kernel virtual address, Va for virtual address, and Pa for physical address. These types are equipped with methods to facilitate address manipulation, conversion between virtual and physical addresses, and safe arithmetic operations on addresses.

In this abstraction, the kernel can seamlessly perform address calculations and fastly convert between physical and kernel virtual adresses.

Both Pa, Va and Kva support arithmetic operations (addition, subtraction, bitwise operations), which allow straightforward address arithmetic. The types can be safely cast between virtual and physical addresses, and the operations maintain the integrity of the memory model.

Arithmetic Operations

Both Pa and Kva types implement various arithmetic operations such as addition, subtraction, and bitwise operations. These operations allow for easy manipulation of addresses, such as incrementing or decrementing by a specific number of bytes or performing logical operations on addresses.

Example Usage:

// Create a physical address
let pa = Pa::new(0x1234_5678_9abc_0000).unwrap();

// Convert to kernel virtual address
let kva = pa.into_kva();

// Perform arithmetic on addresses
let next_pa = pa + 0x1000; // Move to the next page

Structs

Kva

Represents a kernel virtual address.

Pa

Represents a physical address.

Va

Represents a virtual address.

Constants

PAGE_MASK

A mask for extracting the offset within a page from a given address.

PAGE_SHIFT

The shift amount to get the page index from a given address.

PAGE_SIZE

The size of a single page in memory, in bytes.