keos::channel

Struct Sender

pub struct Sender<T: Send + 'static> {
    inner: *mut ChannelInner<T>,
}

The sending-half of channel type.

This half can only be owned by one thread, but it can be cloned to send to other threads.

Messages can be sent through this channel with send.

Fields
inner: *mut ChannelInner<T>
Implementations

impl<T: Send + 'static> Sender<T>

fn inner<'a>(&self) -> &'a ChannelInner<T>

pub fn can_send(&self) -> bool

Can send a value through this channel.

pub fn has_receiver(&self) -> bool

Does anyone can receive a value through this channel.

pub fn send(&self, t: T) -> Result<(), SendError<T>>

Sends a value on this channel.

This function will block until space in the internal buffer becomes available or a receiver is available to hand off the message to.

Note that a successful send does not guarantee that the receiver will ever see the data if there is a buffer on this channel. Items may be enqueued in the internal buffer for the receiver to receive at a later time. If the buffer size is 0, however, the channel becomes a rendezvous channel and it guarantees that the receiver has indeed received the data if this function returns success.

This function will never panic, but it may return Err if the Receiver has disconnected and is no longer able to receive information.

pub fn try_send(&self, t: T) -> Result<(), TrySendError<T>>

Attempts to send a value on this channel without blocking.

This method differs from send by returning immediately if the channelโs buffer is full or no receiver is waiting to acquire some data. Compared with send, this function has two failure cases instead of one (one for disconnection, one for a full buffer).

See send for notes about guarantees of whether the receiver has received the data or not if this function is successful.

pub fn capacity(&self) -> usize

Get capacity of the channel.