keos

Module fs

Filesystem abstraction.

Modules

traits

Defines traits for file system operations.

Structs

Directory

A handle to a directory.

Disk

The disk, a device that has byte sink.

FileBlockNumber

Represents a file block number within a file.

FileSystem

A global file system abstraction.

InodeNumber

Represents a unique identifier for an inode in the filesystem.

RegularFile

A handle to a regular file.

Sector

Sector, an access granuality for the disk.

Enums

File

Represents a file system entry, which can be either a file or a directory.

Statics

FS ๐