keos::fs

Enum File

pub enum File {
    RegularFile(RegularFile),
    Directory(Directory),
}

Represents a file system entry, which can be either a file or a directory.

This enum allows distinguishing between regular files and directories within the filesystem. It provides flexibility for handling different file system objects in a unified manner.

Variants

RegularFile(RegularFile)

A regular file.

This variant represents a standard file in the filesystem, which can be read from or written to.

Directory(Directory)

A directory.

This variant represents a directory in the filesystem, which can contain other files or directories.

Implementations

impl File

pub fn into_regular_file(self) -> Option<RegularFile>

Converts the File into a RegularFile, if it is one.

Returns

Some(RegularFile) if self is a RegularFile.

None if self is not a RegularFile.

This function allows extracting the RegularFile from File safely.

pub fn into_directory(self) -> Option<Directory>

Converts the File into a Directory, if it is one.

Returns

Some(Directory) if self is a Directory.

None if self is not a Directory.

This function allows extracting the Directory from File safely.

Example

let dir = File::Directory(directory);
assert!(dir.into_directory().is_some());

let file = File::RegularFile(regular_file);
assert!(file.into_directory().is_none());

pub fn ino(&self) -> InodeNumber

Get InodeNumber of this File regardless of its inner type.

pub fn size(&self) -> u64

Get size of this File regardless of its inner type.