keos::fs

Struct Directory

pub struct Directory(pub Arc<dyn Directory>);

A handle to a directory.

This struct represents a reference-counted directory that supports file entry management, including opening and removing entries.

Tuple Fields
0: Arc<dyn Directory>
Implementations

impl Directory

pub fn ino(&self) -> InodeNumber

Inode number of the directory.

pub fn size(&self) -> usize

Returns the size of the file in bytes.

pub fn link_count(&self) -> usize

Link count of the directory.

pub fn new(r: impl Directory + 'static) -> Self

Creates a new Directory handle from a given implementation of traits::Directory.

This function takes an instance of any type that implements the traits::Directory trait, wraps it in a reference-counted [Arc], and returns a Directory handle.

Parameters

r: An instance of a type that implements traits::Directory.

Returns

A Directory handle that enables reference-counted access to the underlying file.

pub fn open(&self, path: &str) -> Result<File, KernelError>

Opens a path from the directory.

Parameters

path: The path to the entry.

Returns

Ok(File): The type of the file (e.g., regular file, directory).

Err(Error): An error if the entry cannot be found or accessed.

pub fn create(&self, path: &str, is_dir: bool) -> Result<File, KernelError>

Create an entry in the directory.

Parameters

path: The path to the entry.

is_dir: Indicate whether the entry is directory or not.

Returns

Ok(()): If the entry was successfully added.

Err(Error): An error if the add fails.

pub fn unlink(&self, path: &str) -> Result<(), KernelError>

Unlink an entry in the directory.

Parameters

path: The path to the entry.

Returns

Ok(()): If the entry was successfully added.

Err(Error): An error if the add fails.

pub fn read_dir(&self) -> Result<Vec<(InodeNumber, String)>, KernelError>

Reads the contents of the directory.

This function lists all the entries within the directory.

Returns

Ok(()): If the directory was successfully read.

Err(Error): An error if the read operation fails.

pub fn removed(&self) -> Result<&AtomicBool, KernelError>

Returns AtomicBool which contains whether directory is removed.

This is important because directory operations against the removed directory will result in undesirable behavior (e.g. unreachable file).

Returns

Ok(()): If the directory was successfully read.

Err(Error): An error if the operation fails.