keos::fs

Struct FileBlockNumber

#[repr(transparent)]
pub struct FileBlockNumber(pub usize);

Represents a file block number within a file.

This number refers to the position of a block within a specific file. Each block contains 4096 bytes of contents. It helps in translating file-relative offsets to actual storage locations.

Tuple Fields
0: usize
Implementations

impl FileBlockNumber

pub const fn from_offset(offset: usize) -> Self

Computes the file block number from a byte offset within a file.

In KeOS, each file is divided into file blocks of 0x1000 bytes (4 KiB). This function calculates the file block index corresponding to a given byte offset.

Parameters

offset: The byte offset within the file.

Returns

The FileBlockNumber that corresponds to the given offset.