keos::fs

Struct FileSystem

pub struct FileSystem {
    _p: (),
}

A global file system abstraction.

The FileSystem struct provides an interface for interacting with the file system, including operations such as opening and creating files.

Example

let fs = file_system();
if let Some(file) = fs.open("example.txt") {
    println!("Opened file: {}", file.name());
}

Fields
_p: ()
Implementations

impl FileSystem

pub fn root() -> Directory

Retrieves the root directory of the filesystem.

Returns

Directory: A reference to the root directory.

pub fn register(fs: impl FileSystem + 'static)

Register the global file system.