keos::lang::slab::slab_allocator

Struct SlabAllocator

pub struct SlabAllocator<const BSIZE: usize, const GROW_SIZE: usize> {
    head: AtomicU128,
    stamp: AtomicU64,
    grow_aux: SpinLock<()>,
}

The slab allocator.

Fields
head: AtomicU128stamp: AtomicU64grow_aux: SpinLock<()>
Implementations

impl<const BSIZE: usize, const GROW_SIZE: usize> SlabAllocator<BSIZE, GROW_SIZE>

const REDZONE_SIZE: usize = BSIZE

const BLOCK_SIZE: usize

const G_SIZE: usize

pub(super) const fn new() -> Self

Create a new slab allocator with slab size T.

pub(super) unsafe fn grow(&self, allocator: &Palloc) -> Result<(), AllocError>

Grow the internal blocks by allocating from the physical frame.

pub(super) unsafe fn dealloc(&self, ptr: usize, _allocator: &Palloc)

Deallocate the Block.

pub(super) unsafe fn alloc( &self, allocator: &Palloc, ) -> Result<NonNull<[u8]>, AllocError>

Allocate a Block from the allocator.