keos::mm::page_table

Struct PageTableRoot

#[repr(align(4096))]
pub struct PageTableRoot([Pml4e; 512]);

A page table root.

It wraps the Pml4e array to ensure page table align to 4096. Note that it is not allowed to modified the larger indices than Self::KBASE, which are reserved for kernel address.

Tuple Fields
0: [Pml4e; 512]
Implementations

impl PageTableRoot

pub const KBASE: usize = 256usize

Base of pml4 index occupied for kernel address.

pub fn new_boxed() -> Box<Self>

Create a empty PageTableRoot.

pub fn new_boxed_with_kernel_addr() -> Box<Self>

Create a new PageTableRoot that allowed to access the kernel addresses.

pub fn pa(&self) -> Pa

Get the physical address of this page table root.