keos::mm::page_table

Struct Pdpe

#[repr(transparent)]
pub struct Pdpe(pub usize);

Page Directory Pointer Table Entry (PDPE).

This struct represents a Page Directory Pointer Table Entry (PDPE), the entry of second-level table, in the 4-level page table system for x86_64 architecture. A PDPE is the second-level entry in the virtual memory hierarchy, directly pointing to a Page Directory (PDE) or a higher-level page table that contains further mappings for virtual to physical memory.

The Pdpe struct provides methods for working with the physical address and flags associated with a PDPE, allowing the manipulation of page tables in the virtual memory system.

Tuple Fields
0: usize
Implementations

impl Pdpe

pub const fn pa(&self) -> Option<Pa>

Get the physical address pointed to by this entry.

This function checks whether the page directory pointer table entry is present (i.e., if the โPโ flag is set in the entry). If the entry is present, it extracts the physical address by clearing the flags from the entry.

Returns

Some(Pa) if the PDPE is present, containing the physical address.

None if the PDPE is not present (i.e., the โPโ flag is not set).

pub const fn flags(&self) -> PdpeFlags

Get the flags associated with this entry.

This function extracts the flags from the PDPE, which may indicate whether the page directory pointer table entry is present, writable, user-accessible, etc.

Returns

A PdpeFlags value representing the flags associated with this entry.

pub fn set_pa(&mut self, pa: Pa) -> Result<&mut Self, PageTableMappingError>

Set the physical address for this entry.

This method updates the physical address of the PDPE while preserving the current flags (e.g., read/write permissions). It ensures that the provided physical address is aligned to a 4K boundary (the page size), as required by the architecture.

Parameters

pa: The new physical address to set for the entry.

Returns

Ok(&mut Self) if the address is valid and the update is successful.

Err(PageTableMappingError::Unaligned) if the provided physical address is not aligned.

Warning

This operation does not modify the flags of the entry.

pub fn set_flags(&mut self, perm: PdpeFlags) -> &mut Self

Set the flags for this entry.

This method allows you to update the flags associated with the PDPE without modifying the physical address. It combines the current physical address with the new flags and sets the updated value back into the entry.

Parameters

perm: The new set of flags to assign to the entry.

Returns

A mutable reference to self, allowing for method chaining.

pub fn clear(&mut self) -> Option<Pa>

Clears the entry.

This method removes any previously set physical address and flags from the entry. If the entry contained a valid physical address before being cleared, that address is returned.

Returns

Some(Pa): The physical address that was previously stored in the entry, if it existed.

None: If the entry did not contain a valid physical address.

pub fn into_pd_mut(&mut self) -> Result<&mut [Pde], PageTableMappingError>

Get a mutable reference to the page directory pointed to by this entry.

This method retrieves a mutable reference to the page directory that this PDPE points to, assuming that the entry is present (i.e., the โPโ flag is set).

Returns

Ok(&mut [Pde]) if the page directory is valid, represented as a mutable slice of Pde (page directory entries).

Err(PageTableMappingError::NotExist) if the PDPE is not present or invalid.

Safety

This operation assumes that the physical address of the page directory is valid and properly aligned.

pub fn into_pd(&self) -> Result<&[Pde], PageTableMappingError>

Get a reference to the page directory pointed to by this entry.

This method retrieves an immutable reference to the page directory that this PDPE points to, assuming that the entry is present (i.e., the โPโ flag is set).

Returns

Ok(&[Pde]) if the page directory is valid, represented as an immutable slice of Pde (page directory entries).

Err(PageTableMappingError::NotExist) if the PDPE is not present or invalid.

Safety

This operation assumes that the physical address of the page directory is valid and properly aligned.