keos::mm::page_table

Struct StaleTLBEntry

pub struct StaleTLBEntry(Va, Page);

Struct for invalidating the TLB (Translation Lookaside Buffer) entry.

This struct is responsible for invalidating a TLB entry associated with a specific virtual address (Va). TLB entries are cached mappings between virtual addresses and physical addresses, and they need to be invalidated when the corresponding page table entries are modified or removed.

This struct provides methods for invalidating the TLB entry and safely forgetting the modification. This internally holds the page to be invalidated, to delay the free until the tlb entry is invalidated.

Tuple Fields
0: Va1: Page
Implementations

impl StaleTLBEntry

pub fn new(va: Va, page: Page) -> Self

Create a new StaleTLBEntry.

pub fn invalidate(self) -> Page

Invalidate the underlying virtual address.

This method issues an assembly instruction to invalidate the TLB entry corresponding to the given virtual address. The invalidation ensures that any cached translations are cleared and that the system will use the updated page table entries for subsequent address lookups.