keos::mm

Struct ContigPages

pub struct ContigPages {
    arena_idx: usize,
    kva: Kva,
    cnt: usize,
    ref_cnt: &'static AtomicU64,
}

A contiguous pages representation.

Fields
arena_idx: usizekva: Kvacnt: usizeref_cnt: &'static AtomicU64
Implementations

impl ContigPages

pub fn new(size: usize) -> Option<Self>

Allocate a page.

pub fn new_with_align(size: usize, align: usize) -> Option<Self>

Allocate a page with align

pub fn kva(&self) -> Kva

Get virtual address of this page.

pub unsafe fn from_va(kva: Kva, size: usize) -> Self

Constructs a page from a kva.

Safety

This function is unsafe because improper use may lead to memory problems. For example, a double-free may occur if the function is called twice on the same raw pointer.

pub fn split(self) -> Vec<Page>

Split the ContigPages into multiple pages.