keos::mm

Struct Page

pub struct Page {
    inner: ContigPages,
}

A representation of a memory page.

The Page struct encapsulates a single memory page, providing methods to allocate, access, and manipulate the underlying pageโs contents.

This page internally holds the reference counts. This counter increases on a calling of Page::clone, and decreases when the page instance is dropped.

Example:

let page = Page::new().unwrap();
let va = page.va();  // Get the virtual address of the page.
let pa = page.pa();  // Get the physical address of the page.

Fields
inner: ContigPages
Implementations

impl Page

pub fn new() -> Self

Allocate a new page.

This function allocates a new memory page.

pub fn kva(&self) -> Kva

Get the kernel virtual address of this page.

Returns

The kernel virtual address (Kva) of the page.

pub fn pa(&self) -> Pa

Get the physical address of this page.

Returns

The physical address (Pa) of the page.

pub fn into_raw(self) -> Pa

Consumes the page, returning its physical address.

This method โconsumesโ the Page and returns its physical address. After calling this function, the caller is responsible for managing the memory previously associated with the page. It is important to properly release the page, which can be done using Page::from_pa.

Returns

The physical address (Pa) of the page.

pub unsafe fn from_pa(pa: Pa) -> Self

Constructs a page from a given physical address.

This method reconstructs a Page from a physical address (Pa). It should be used only after consuming a Page with Page::into_raw. The physical address passed must be valid.

Safety

This function is unsafe because incorrect usage could result in memory issues such as a double-free.

Ensure that the physical address passed is valid and is being used correctly.

Arguments

pa: The physical address of the page.

Returns

A Page reconstructed from the physical address.

pub fn inner(&self) -> &[u8]

Get a reference to the underlying slice of the page (read-only).

This method allows access to the contents of the page as a byte slice. The caller can read from the pageโs memory, but cannot modify it.

Returns

A reference to the byte slice representing the contents of the page.

pub fn inner_mut(&mut self) -> &mut [u8]

Get a mutable reference to the underlying slice of the page.

This method allows modification of the contents of the page as a byte slice. The caller can read from and write to the pageโs memory.

Returns

A mutable reference to the byte slice representing the contents of the page.