keos::sync::rwlock

Struct RwLockWriteGuard

pub struct RwLockWriteGuard<'a, T>
where
    T: ?Sized + Send + 'a,
{
    lock: &'a RwLock<T>,
    data: &'a mut T,
}

RAII structure used to release the exclusive write access of a lock when dropped.

This structure is created by the write and try_write methods on RwLock.

Fields
lock: &'a RwLock<T>data: &'a mut T
Implementations

impl<'a, T> RwLockWriteGuard<'a, T>
where T: ?Sized + Send + 'a,

pub fn downgrade(self) -> RwLockReadGuard<'a, T>

Downgrade the RwLockWriteGuard into RwLockReadGuard.