keos::syscall::uaccess

Struct UserU8SliceRO

pub struct UserU8SliceRO {
    addr: usize,
    len: usize,
}

A one-time, read-only pointer to a slice of u8 in user-space.

This struct allows the kernel to safely read from a user-space buffer while preventing TOCTOU attacks by taking ownership of the pointer during operations.

Fields
addr: usizelen: usize
Implementations

impl UserU8SliceRO

pub fn new(addr: usize, len: usize) -> Self

Creates a new UserU8SliceRO instance with the given user-space address and length.

pub fn get(self) -> Result<Vec<u8>, KernelError>

Reads data from the user-space buffer into a Vec<u8>.

Takes ownership of self to prevent TOCTOU attacks.

Returns Ok(Vec<u8>) containing the data if successful, otherwise Err(KernelError::BadAddress).