keos::syscall::uaccess

Struct UserU8SliceWO

pub struct UserU8SliceWO {
    addr: usize,
    len: usize,
}

A one-time, write-only pointer to a slice of u8 in user-space.

This struct allows the kernel to safely write to a user-space buffer while preventing TOCTOU attacks by taking ownership of the pointer during operations.

Fields
addr: usizelen: usize
Implementations

impl UserU8SliceWO

pub fn new(addr: usize, len: usize) -> Self

Creates a new UserU8SliceWO instance with the given user-space address and length.

pub fn put(self, other: &[u8]) -> Result<usize, KernelError>

Writes data from a slice to the user-space buffer.

Takes ownership of self to prevent TOCTOU attacks.

Returns Ok(usize) indicating the number of bytes written, or Err(KernelError::BadAddress) on failure.