keos::teletype

Struct Serial

pub struct Serial {
    _p: (),
}

A serial teletype interface for x86_64 systems.

This struct provides a basic implementation of a serial TTY using the COM1 serial port. It implements the Teletype trait to allow read and write operations over a serial interface.

Fields
_p: ()
Implementations

impl Serial

pub const fn new() -> Self

Creates a new COM1 serial interface instance.

This function initializes a serial TTY for performing character-based I/O operations. The actual hardware interaction is handled via the Teletype trait methods (write and read).

Returns

A new instance of Serial, representing a COM1 serial interface.