keos::thread

Struct Current

pub struct Current {
    _p: (),
}

The opaque structure indicating the running thread on the current cpu.

Fields
_p: ()
Implementations

impl Current

pub fn park_with(f: impl FnOnce(ParkHandle))

Run a function f with ParkHandle for current thread, and then park the current thread.

pub fn exit(exit_code: i32) -> !

Exit the current thread with exit_code.

pub fn get_tid() -> u64

Get the current thread’s id.