keos

Module util

Debugging Utilities.

Functions

copy_file

Copy a RegularFile’s content into another RegularFile.

hex_dump⚠

Dumps the memory occupied by a value of type T to the console.

hex_dump_slice

Dumps the bytes in buf to the console as hex bytes, arranged 16 per line.